+++ src/app/controllers/eventos.js
@@ -1,3 +1,4 @@
+const { count } = require('./../models/eventos')
 const EventoSchema = require('./../models/eventos')
 
 class Evento {
@@ -14,15 +15,32 @@ class Evento {
     })
   }
 
+  getByTitle(req, res) {
+    let title = req.params.title.replace(/%20/g, " ")
+
+    EventoSchema.findOne({ titulo: title }, (err, data) => {
+      if (err) {
+        res.status(500).json({ message: 'Houve um erro ao processar sua requisição', error: err })
+      } else {
+        res.status(200).json({ message: 'Evento recuperado com sucesso', data: data })
+      }
+    })
+  }
+
   getWithParams(req, res) {
-    let limit = parseInt(req.query.limit)
+    const limit = 10
     let query = {}
     let page = req.query.page
     let skip = limit * (page - 1)
-    let {
-      columnSort,
-      valueSort
-    } = req.query
+    let { keyword, category, columnSort, valueSort } = req.query
+
+    if (category) {
+      query['tipo'] = new RegExp(category, "i")
+    }
+
+    if (keyword) {
+      query = { $text: { $search: `"\"${keyword}\""` } }
+    }
 
     EventoSchema
       .find(query)
@@ -33,10 +51,25 @@ class Evento {
         if (err) {
           res.status(500).json({ message: 'Houve um erro ao processar sua requisição', error: err })
         } else {
-          res.status(200).json({ message: 'Dados recuperados com sucesso', data: data, page: page, limit: limit })
+          EventoSchema
+            .estimatedDocumentCount()
+            .find(query)
+            .exec((err, count) => {
+              let totalDocuments = count.length
+              if (err) {
+                res.status(500).json({ message: 'Houve um erro ao processor sua requisição', error: err })
+              } else {
+                if (totalDocuments > 0) {
+                  res.status(200).json({ message: 'Dados recuperados com sucesso', data: data, page: page, limit: limit, count: totalDocuments })
+                } else {
+                  res.status(204).json({ message: 'Não há dados para serem exibidos' })
+                }
+              }
+            })
         }
       })
   }
+  
 }
 
 module.exports = new Evento()